better c-archive separation
